misc: Added LR_Q and SC_Q instructions for 64-byte data
This commit consists of all the files required to be modified
to implement two micro-ops LR_Q and SC_Q, but this is using a
older version, we need to use a newer one, or the latest one.

Besides, in order to implement ENQCMD, we need to make sure
that ENQCMD itself is a macro-op, which contains two static
instruction pointers: LR_Q and SC_Q. Refer to
src/arch/riscv/isa/formats/amo.isa for more details.

Also, in this implementation, we implement a new architectural
state register, which is not appropriate. ENQCMD does load and
store, either both succeeded or failed. So we need to make sure
the 512-bit register is not visible to other instructions, i.e.
we need to add more registers (fake registers) which are not
architectural registers (those 32 registers), to store the
temporary 512-bit value.

The current implementation of SC_Q does not work using simple
inline assembly, since store conditional requires a "reserved"
address and we need to reserve such address before SC_Q is
executed. So if we enable the debug-flags IntRegs, then we will
see that "Addr reserved? no" being printed out many times.

Hence, the next step is to git pull the latest version of gem5,
and create a new branch, to implement macro-op ENQCMD.

Change-Id: I224fdb50eaafa7df7e5850efc9f3df93cfbfaccd